v1.9.6 Update || The long awaited update.
Hello, its has been a long time hasn’t it. Almost a year now since the
last update. I am very sorry for not updating this to 1.10+ sooner but
I didn’t have time to try and find another particle API that was
updated and then rewrite my code to allow that new API. But thanks to
@PazuHD he sent me an updated version of the ParticleAPI I use. So
without a further ado, here is the new update that added the long
wanted 1.10+ support.

Bug Fixes:
  - Some servers get an error from the ItemMergeEvent and so it was
removed and instead of using that the fountain items just have a random
name to avoid merging.
  - Fixed the food and Pokemon fountains so they don't have random
heads in it.

Added:
  - Added support for versions 1.10 - 1.11.2.
  - A few new for the food fountain.
  - New MCUpdate stats to view the usage of Block Particles. Link:
http://mcupdate.org/plugin?a=metrics&pl=75